requirements.txt
bertopic[flair]==0.17.3
sentence-transformers==5.1.0
faiss-cpu==1.11.0
huggingface_hub==0.34.4
datasets==3.6.0
numpy==1.26.4
python-dotenv==1.1.1
ipywidgets==8.1.7
streamlit==1.48.0
torch==2.6.0
boto3==1.40.8
jupyterlab==4.4.6